refactor: extract StopBrowser from BottomSheet
Insert a StopBrowser layer between BottomSheet and its content:

  BottomSheet                 BottomSheet
  ├─ grab handle      ->      ├─ grab handle
  ├─ BottomSheetHeader        └─ StopBrowser
  └─ StopGrid                    ├─ BottomSheetHeader
                                 └─ StopGrid

StopBrowser is the shell-agnostic nearby-stop browsing surface (header
+ grid). It owns the surface-local filter state (view / agency /
route_type) and the stopTimes trimming. BottomSheet keeps only the
shell concerns — positioning, grab handle, drag-to-expand — and hosts
a StopBrowser as its content.

This lets StopBrowser be reused in other shells (the multi-pane panel,
a future sidebar) without going through BottomSheet. The collapse-on-
select behaviour stays wired in BottomSheet, which passes a wrapped
onStopSelected to StopBrowser.
